input file custom component refactor
