+++ packages/react-ui/storybook/.storybook/preview.js
@@ -47,7 +47,7 @@ export const parameters = {
         'CARTO Theme',
         ['Palette', 'Typography'],
         'Common',
-        'Components',
+        'Custom Components',
         'Widgets',
         ['CategoryWidget', 'FormulaWidget', , 'HistogramWidget', 'PieWidget'],
         'Widgets UI',
