+++ packages/react-ui/storybook/stories/custom-components/InputFile.stories.js
@@ -1,0 +1,137 @@
+import {
+  Box,
+  Button,
+  CircularProgress,
+  FormControl,
+  InputLabel,
+  Typography
+} from '@material-ui/core';
+import React, { useState } from 'react';
+import InputFile from '../../../src/custom-components/InputFile';
+import { Alert } from '@material-ui/lab';
+import UploadIcon from '../../../src/assets/UploadIcon';
+
+const options = {
+  title: 'Custom Components/InputFile',
+  component: InputFile,
+  argTypes: {
+    placeholder: {
+      defaultValue: 'Drag and drop your file or click to browse',
+      description: 'Text to display as input placeholder.',
+      control: {
+        type: 'text'
+      }
+    },
+    buttonText: {
+      defaultValue: 'Browse',
+      description: 'Text to display on the browse button.'
+    },
+    files: {
+      defaultValue: [],
+      description: 'Files handled by the input.'
+    },
+    accept: {
+      defaultValue: 'application/JSON',
+      description: 'String that defines the file types the file input should accept.',
+      control: {
+        type: 'string'
+      }
+    },
+    multiple: {
+      defaultValue: false,
+      description: 'Specifies that multiple files can be chosen at once.',
+      control: {
+        type: 'boolean'
+      }
+    }
+  }
+};
+export default options;
+
+const Template = ({ ...args }) => <InputFile {...args} />;
+
+const ExampleTemplate = ({ ...args }) => {
+  const [files, setFiles] = useState([]);
+  const [uploading, setUploading] = useState(false);
+  const [alert, setAlert] = useState();
+
+  const handleInputFileChange = (files) => {
+    // TODO: validate files
+    setAlert({
+      severity: 'warning',
+      message: 'Some files are too big. The upload will take some time.',
+      actionText: 'Close'
+    });
+
+    setFiles(files);
+  };
+
+  const handleUpload = () => {
+    setUploading(true);
+
+    setTimeout(() => {
+      setUploading(false);
+      setFiles([]);
+
+      setAlert({
+        severity: 'error',
+        message: 'Oops! We seem to have a problem with the format of your dataset.',
+        actionText: 'Close'
+      });
+    }, 2000);
+  };
+
+  return (
+    <div>
+      <FormControl size='small' margin='dense'>
+        <InputLabel>Import data</InputLabel>
+        <Box
+          display='flex'
+          flexDirection='row'
+          justifyContent='space-between'
+          alignItems='center'
+        >
+          <InputFile {...args} files={files} onChange={handleInputFileChange} />
+          <Button
+            color='primary'
+            size='small'
+            startIcon={uploading ? <CircularProgress size={16} /> : <UploadIcon />}
+            disabled={!files.length}
+            onClick={handleUpload}
+          >
+            Upload
+          </Button>
+        </Box>
+        <Box mt={1}>
+          <Typography
+            variant='caption'
+            color='textSecondary'
+            style={{ fontWeight: 'normal' }}
+          >
+            Upload a CSV or GeoJSON file, or a zip package with your Shapefile
+          </Typography>
+        </Box>
+      </FormControl>
+
+      {alert && (
+        <Box mt={1}>
+          <Alert severity={alert.severity} onClose={() => setAlert(null)}>
+            {alert.message}
+          </Alert>
+        </Box>
+      )}
+    </div>
+  );
+};
+
+export const Default = Template.bind({});
+Default.args = {
+  accept: 'image/*',
+  multiple: true
+};
+
+export const Example = ExampleTemplate.bind({});
+Example.args = {
+  accept: 'image/*',
+  multiple: true
+};
